Restructuring code to match go conventions addressing #14
