+++ Makefile
@@ -1,28 +1,51 @@
 CGO_ENABLED=0
 TRAVIS_BUILD_NUMBER?=debug0
 
+AWSNYCAST := AWSnycast
+SRCDIR         := src
+
 .PHONY: coverage get test clean
 
 all: get coverage AWSnycast
 
-AWSnycast: *.go */*.go
-	go build -a -tags netgo -ldflags '-w' .
-	strip AWSnycast
+PKGS           := \
+	$(AWSNYCAST)/\
+	$(AWSNYCAST)/aws \
+	$(AWSNYCAST)/config \
+	$(AWSNYCAST)/daemon \
+	$(AWSNYCAST)/healthcheck \
+	$(AWSNYCAST)/instancemetadata \
+	$(AWSNYCAST)/utils
+
+SOURCES        := $(foreach pkg, $(PKGS), $(wildcard $(SRCDIR)/$(pkg)/*.go))
+
+GOPATH  := $(shell pwd -L)
+export GOPATH
+
+PATH := bin:$(PATH)
+export PATH
+
+AWSnycast: $(SOURCES)
+	@echo Building $(AWSNYCAST)...
+	bin/gom build -a -tags netgo -ldflags '-w' -o bin/$(AWSNYCAST) $@
+	strip bin/$(AWSNYCAST)
 
 test:
-	go test -short ./...
+	bin/gom test -short ./...
 
 get:
-	CGO_ENABLED=0 go get -a -x -installsuffix cgo -ldflags '-d -s -w' && go install -a -x -installsuffix cgo -ldflags '-d -s -w'
+	@echo Getting dependencies...
+	@go get github.com/mattn/gom
+	@bin/gom install
 
 fmt:
-	go fmt ./...
+	bin/gom fmt ./...
 
 coverage:
-	go test -cover -short ./...
+	bin/gom test -cover -short ./...
 
 integration:
-	go test ./...
+	bin/gom test ./...
 
 clean:
 	rm -rf dist */coverage.out */coverprofile.out coverage.out coverprofile.out AWSnycast
@@ -31,11 +54,7 @@ realclean: clean
 	make -C tests/integration realclean
 
 coverage.out:
-	cd aws ; go test -coverprofile=coverage.out ./... ; cd ..
-	cd daemon ; go test -coverprofile=coverage.out ./... ; cd ..
-	cd config ; go test -coverprofile=coverage.out ./... ; cd ..
-	cd healthcheck ; go test -coverprofile=coverage.out ./... ; cd ..
-	cd instancemetadata ; go test -coverprofile=coverage.out ./... ; cd ..
+	@$(foreach pkg, $(PKGS), bin/gom test -coverprofile=coverage.out $(pkg);)
 	echo "mode: set" > coverage.out && cat */coverage.out | grep -v mode: | sort -r | awk '{if($$1 != last) {print $$0;last=$$1}}' >> coverage.out
 
 itest_%:
