+++ package/Makefile
@@ -27,7 +27,7 @@ $(PACKAGE_FILE): .docker_container_id
 
 .docker_container_id: .docker_image_id
 	docker run --rm=false \
-		-v $(CURDIR)/..:/go/src/github.com/bobtfish/AWSnycast:rw \
+		-v $(CURDIR)/../src/AWSnycast:/go/src/github.com/bobtfish/AWSnycast:rw \
 		-v $(CURDIR)/build.sh:/build.sh:ro \
 		--cidfile=$(CURDIR)/.docker_container_id \
 		$$(cat .docker_image_id) \
