+++ src/AWSnycast/aws/manage_route_spec.go
@@ -1,12 +1,12 @@
 package aws
 
 import (
+	"AWSnycast/healthcheck"
+	"AWSnycast/instancemetadata"
 	"errors"
 	"fmt"
 	log "github.com/Sirupsen/logrus"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/bobtfish/AWSnycast/healthcheck"
-	"github.com/bobtfish/AWSnycast/instancemetadata"
 	"github.com/hashicorp/go-multierror"
 	"net"
 	"strings"
