+++ src/AWSnycast/config/config_test.go
@@ -1,14 +1,14 @@
 package config
 
 import (
+	"AWSnycast/aws"
+	"AWSnycast/healthcheck"
+	"AWSnycast/instancemetadata"
+	"AWSnycast/testhelpers"
 	"errors"
 	"fmt"
 	a "github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/bobtfish/AWSnycast/aws"
-	"github.com/bobtfish/AWSnycast/healthcheck"
-	"github.com/bobtfish/AWSnycast/instancemetadata"
-	"github.com/bobtfish/AWSnycast/testhelpers"
 	"github.com/hashicorp/go-multierror"
 	"github.com/stretchr/testify/assert"
 	"testing"
