+++ src/AWSnycast/instancemetadata/instancemetadata_test.go
