+++ src/AWSnycast/tests/integration/getvariables.rb
